requirements.txt
Django==5.0.2
django-ckeditor==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
Pillow==10.2.0
reportlab===4.1.0